separation of logic for passport and router
